refactor(backend): Split utils/task.rs into dedicated modules for every config type
