+++ backend/migration/src/lib.rs
@@ -46,6 +46,6 @@ mod m20250503_182326_create_external_team_invitation;
 mod m20250526_105526_add_team_color_field;
 mod m20250526_172108_add_organization_column_to_team;
 mod m20250604_000001_new_identification_system;
+mod m20260412_000001_remove_team_confirmation_fields;
 mod m20260429_225143_add_first_name_to_users;
 mod m20260429_225144_onboarding_field_rework;
-mod m20260412_000001_remove_team_confirmation_fields;
