+++ backend/src/models/mod.rs
@@ -1,10 +1,9 @@
 pub mod ctf_experience;
 pub mod email_verification_request;
-pub mod event_config;
 pub mod external_team_invitation;
 pub mod flag_capture;
 pub mod school_grade;
-pub mod task;
+pub mod task_manager;
 pub mod team_invites;
 pub mod team_status;
 pub mod teams;
