+++ backend/src/models/task_manager/mod.rs
@@ -1,0 +1,4 @@
+pub mod event_config;
+pub mod participant_tags_config;
+pub mod registration_config;
+pub mod task_config;
