+++ backend/src/utils/points_counter.rs
@@ -1,5 +1,5 @@
 use crate::entities::{flag_capture, teams};
-use crate::models::event_config::EventStageType;
+use crate::models::task_manager::event_config::EventStageType;
 use crate::utils::app_state::AppState;
 use crate::utils::error::Error;
 use chrono::NaiveDateTime;
