+++ backend/tests/routes/event/participant_tags.rs
@@ -3,7 +3,7 @@ use crate::test_utils::database::TestDatabase;
 use crate::test_utils::task_manager::create_default_test_task_manager;
 use actix_web::test;
 use actix_web::test::read_body_json;
-use hack4krak_backend::models::task::ParticipantTag;
+use hack4krak_backend::models::task_manager::participant_tags_config::ParticipantTag;
 use hack4krak_backend::services::task_manager::TaskManager;
 
 #[actix_web::test]
